update homepage with analytics overview
